requirements.txt
boto3>=1.34.0
anthropic>=0.18.0
requests>=2.31.0
python-dotenv>=1.0.0
pydantic>=2.5.0
fastapi>=0.104.0
uvicorn>=0.24.0
email-validator>=2.1.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
jinja2>=3.1.0
python-multipart>=0.0.6
